requirements.txt
fastapi[standard]
sqlmodel
python-jose
passlib
bcrypt